+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 1.5.0 (2016-01-01)
 -----------------
 
-  * added alias `bovigo\assert\assertEmpyString()`
-  * added alias `bovigo\assert\assertEmpyArray()`
+  * added alias `bovigo\assert\assertEmptyString()`
+  * added alias `bovigo\assert\assertEmptyArray()`
 
 
 1.4.0 (2015-12-31)